requirements.txt
bcbio-gff==0.6.6

# Requirements below comes from the https://github.com/microbiomedata/nmdc-metadata/requirements.txt

################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

biolinkml >= 1.6.0
click
pipenv-to-requirements
pyyaml
requests

